1、新增了ijkPlayer 2、优化KsgPlayer代码 3、增加ReaderView组件

+++ lib_ksgPlayer/src/main/java/com/ksg/ksgplayer/receiver/PlayerStateGetter.java
@@ -32,14 +32,14 @@ public interface PlayerStateGetter {
      *
      * @return current
      */
-    int getProgress();
+    long getProgress();
 
     /**
      * 获取总时长
      *
      * @return duration
      */
-    int getDuration();
+    long getDuration();
 
     /**
      * 获取缓冲进度.
